[gulpy] Major refactoring: move funcs to other files
